autotests/requirements.txt
pytest
requests
web3
PyYAML
